+++ src/eepp/maps/mapeditor/mapobjectproperties.cpp
@@ -3,7 +3,7 @@
 
 namespace EE { namespace Maps { namespace Private {
 
-static UITextView * createTextBox( const String& Text = "", UINode * Parent = NULL, const Sizei& Size = Sizei(), const Vector2f& Pos = Vector2f(), const Uint32& Flags = UI_CONTROL_DEFAULT_FLAGS | UI_AUTO_SIZE, const Uint32& fontStyle = Text::Regular ) {
+static UITextView * createTextBox( const String& Text = "", UINode * Parent = NULL, const Sizef& Size = Sizef(), const Vector2f& Pos = Vector2f(), const Uint32& Flags = UI_CONTROL_DEFAULT_FLAGS | UI_AUTO_SIZE, const Uint32& fontStyle = Text::Regular ) {
 	UITextView * Ctrl = UITextView::New();
 	Ctrl->setFontStyle( fontStyle );
 	Ctrl->resetFlags( Flags )->setParent( Parent )->setSize( Size )->setVisible( true )->setEnabled( false )->setPosition( Pos );
@@ -36,14 +36,14 @@ MapObjectProperties::MapObjectProperties( GameObjectObject * Obj ) :
 	Int32 InitialY		= 16;
 	Int32 DistFromTitle	= 18;
 
-	UITextView * Txt = createTextBox( "Object name:", mUIWindow->getContainer(), Sizei(), Vector2f( 50, InitialY ), UI_CONTROL_DEFAULT_FLAGS | UI_AUTO_SIZE, Text::Shadow );
+	UITextView * Txt = createTextBox( "Object name:", mUIWindow->getContainer(), Sizef(), Vector2f( 50, InitialY ), UI_CONTROL_DEFAULT_FLAGS | UI_AUTO_SIZE, Text::Shadow );
 	mUIInput = UITextInput::New();
 	mUIInput->setParent( mUIWindow->getContainer() )->setSize( 120, 0 )->setPosition( Txt->getPosition().x + DistFromTitle, Txt->getPosition().y + DistFromTitle );
 	mUIInput->setMaxLength( 64 );
 	mUIInput->setText( mObj->getName() );
 	mUIInput->addEventListener( UIEvent::OnPressEnter, cb::Make1( this, &MapObjectProperties::onOKClick ) );
 
-	UITextView * Txt2 = createTextBox( "Object type:", mUIWindow->getContainer(), Sizei(), Vector2f( 50+192, InitialY ), UI_CONTROL_DEFAULT_FLAGS | UI_AUTO_SIZE, Text::Shadow );
+	UITextView * Txt2 = createTextBox( "Object type:", mUIWindow->getContainer(), Sizef(), Vector2f( 50+192, InitialY ), UI_CONTROL_DEFAULT_FLAGS | UI_AUTO_SIZE, Text::Shadow );
 	mUIInput2 = UITextInput::New();
 	mUIInput2->setParent( mUIWindow->getContainer() )->setSize( 120, 0 )->setPosition( Txt2->getPosition().x + DistFromTitle, Txt2->getPosition().y + DistFromTitle );
 	mUIInput2->setMaxLength( 64 );
@@ -51,8 +51,8 @@ MapObjectProperties::MapObjectProperties( GameObjectObject * Obj ) :
 	mUIInput2->addEventListener( UIEvent::OnPressEnter, cb::Make1( this, &MapObjectProperties::onOKClick ) );
 
 	Uint32 TxtBoxFlags = UI_CONTROL_DEFAULT_FLAGS | UI_HALIGN_CENTER | UI_VALIGN_CENTER;
-	createTextBox( "Property Name", mUIWindow->getContainer(), Sizei(192, 24), Vector2f( 50, mUIInput->getPosition().y + mUIInput->getSize().getHeight() + 12 ), TxtBoxFlags, Text::Shadow );
-	UITextView * TxtBox = createTextBox( "Property Value", mUIWindow->getContainer(), Sizei(192, 24), Vector2f( 50+192, mUIInput->getPosition().y + mUIInput->getSize().getHeight() + 12 ), TxtBoxFlags, Text::Shadow );
+	createTextBox( "Property Name", mUIWindow->getContainer(), Sizef(192, 24), Vector2f( 50, mUIInput->getPosition().y + mUIInput->getSize().getHeight() + 12 ), TxtBoxFlags, Text::Shadow );
+	UITextView * TxtBox = createTextBox( "Property Value", mUIWindow->getContainer(), Sizef(192, 24), Vector2f( 50+192, mUIInput->getPosition().y + mUIInput->getSize().getHeight() + 12 ), TxtBoxFlags, Text::Shadow );
 
 	UIPushButton * OKButton = UIPushButton::New();
 	OKButton->setParent(  mUIWindow->getContainer() )->setSize( 80, 0 );
